mapbox/src/test/java/com/graphhopper/navigation/mapbox/MapboxResourceTest.java
package com.graphhopper.navigation.mapbox;

import org.junit.Test;

import java.util.Collections;
import java.util.List;

import static org.junit.Assert.assertEquals;

public class MapboxResourceTest {


    @Test
    public void voiceInstructionsTest() {

        List<Double> bearings = MapboxResource.getBearing("");
        assertEquals(0, bearings.size());
        assertEquals(Collections.EMPTY_LIST, bearings);

        bearings = MapboxResource.getBearing("100,1");
        assertEquals(1, bearings.size());
        assertEquals(100, bearings.get(0), .1);

        bearings = MapboxResource.getBearing(";100,1;;");
        assertEquals(4, bearings.size());
        assertEquals(100, bearings.get(1), .1);
    }

}
